'/health' endpoint is working
